Move test to regression tests
